+++ fvid/cythonizer.py
@@ -1,6 +1,9 @@
 #cython: language_level = 3
 from distutils.core import Extension, setup
 from Cython.Build import cythonize
+import os
+
+os.environ['CFLAGS'] = "-march='native'"
 
 ext = Extension(name="fvid_cython", sources=["fvid_cython.pyx"])
 setup(ext_modules=cythonize(ext, compiler_directives={'language_level': 3, 'infer_types': True}))