+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 __Planned (1.1.0)__
 - GUI
 - Use json-encoded strings instead of pickle for storing stuff
+- Allow H.265 encoding/decoding with the -5 or --h265 flags. This makes the video file about 50% smaller, and decoding the pixels about 3x faster.
 
 __1.0.0__